requirements.txt
# Project Requirements
# Python >= 3.8

# Core Dependencies
python-dotenv==1.0.0
pydantic==2.0.0

# Database Connectivity
snowflake-connector-python==3.0.0
psycopg2-binary==2.9.0
sqlalchemy==2.0.0

# Data Processing
pandas==2.0.0
pandas-gbq==0.20.0
numpy==1.24.0

# Apache Spark
pyspark==3.5.0
pyarrow==12.0.0

# Web Framework & Visualization
streamlit==1.28.0
plotly==5.17.0
dash==2.0.0
altair==5.0.0

# Encryption & Security
cryptography==41.0.0
bcrypt==4.1.0

# Data Quality & Validation
great-expectations==0.17.0
jsonschema==4.19.0

# Performance & Monitoring
prometheus-client==0.17.0
python-json-logger==2.0.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.0

# Code Quality
pylint==3.0.0
black==23.11.0
flake8==6.1.0

# Utilities
requests==2.31.0
click==8.1.0
tqdm==4.66.0

# Configuration Management
pyyaml==6.0
configparser==6.0.0
